Fix UB bug in extensions/cl_khr_command_buffer/command_buffer_test_barrier.cpp
BarrierWithWaitListKHR::Run in command_buffer_test_barrier.cpp contains a
"check second enqueue" which tries to zero out the cl_mems before enqueueing
the command buffer again. However, the zeroing commands are enqueued to the
in-order queue, while the command buffer to the out-of-order queue. Since there
is no implicit or explicit sync, this is undefined behaviour.