Escape subnormal values
Currently we don't escape subnormal values when generating image data. In
sampler read tests, we use `!=` to check the two values even when it is
floating-point data, which requires the two values are bitwise equal. However, a
sampler might flush subnormal values, causing the test case to fail.

In this patch, when generating random image data, we escape subnormal values.