relationals: fix out of bounds accesses in shuffle test (#1961)
The values in `src` are indices into an array in `get_order_string()`.
Not initializing `src` resulted in out of bounds accesses there.

It seems that when the out of bounds accesses happened, the result of
`get_order_string()` was not actually used, so at least the test was not
using random data. Fix the issue as it prevents a clean run of this test
with e.g. AddressSanitizer.

Signed-off-by: Sven van Haastregt <sven.vanhaastregt@arm.com>