Skip `RunPropArrayInfoTest` if there are no properties
When there are no properties, the value passed as `param_value_size` is
-1, but since the argument is of type `size_t`, this is interpreted as a
really big value (and hence not an invalid value, as the test intends).

Moreover, if the `command_buffer` was created without properties, then
even a size of 0 is large enough to store them. This test requires
`command_buffer` to have properties for there to be invalid values for
`param_value_size`.

Signed-off-by: Gorazd Sumkovski <gorazd.sumkovski@arm.com>