Fix payload value query for external semaphores
When creating a CL semaphore object from a Vulkan semaphore one, we
explicitly pass `-1` as the file descriptor value in the case of
`VULKAN_EXTERNAL_SEMAPHORE_HANDLE_TYPE_SYNC_FD`. According to the CL
specification:

    The special value -1 for fd is treated like a valid sync file
    descriptor referring to an object that has already signaled. The
    import operation will succeed and the semaphore will have a
    temporarily imported payload as if a valid file descriptor had
    been provided.

The test currently checks that the semaphore payload is unsignalled,
unconditionally, which is incorrect.

Changed the test to check for the correct expected payload value.

Signed-off-by: Ahmed Hesham <ahmed.hesham@arm.com>