fix workgroup sizes in mutable_command_global_size test
test MutableDispatchGlobalSize was failing (with an implementation that
doesn't support non-uniform WGs), because the update_global_size = 3
was not a multiple of the local work-size.

Fixed by increasing the global work-size to 256K and update size to 16K.
This should work with all devices that have max_work_group_size <= 16K.

Updates also MutableDispatchWorkGroups which had an out-of-bounds access,
because it hardcoded the value of global work-size.