Use the system package manager to install cross-compilers
On Ubuntu, use the system's package manager, `apt`, to download and install
the cross-compilers for Arm and AArch64. This replaces downloading the
compilers as tarballs.

To ensure that the correct version of the compiler is used when calling one
without the version suffix, e.g. `aarch64-linux-gnu-gcc`, use
`update-alternatives`.

On Linux, the compilers for `x86_64`, `arm`, and `aarch64` will now correctly
use the filesystem root as their sysroots, i.e. `/usr/include` will be used
when searching for headers, instead of each compiler providing its own.