Add support for environment variable override
Add the ability to specify the type of DMA heap to use when allocating
a DMA buffer, the path of which can be overridden using an environment
variable, `OCL_CTS_DMA_HEAP_PATH_SYSTEM`.

Add a new helper function to check the validity of a DMA buffer
handle (file descriptor).

Add Doxygen-style documentation for the new functions.

Signed-off-by: Ahmed Hesham <ahmed.hesham@arm.com>