Check command-buffer extension version for equality
As described in Issue https://github.com/KhronosGroup/OpenCL-CTS/issues/2152
we currently check the provisional extension version supported by a vendor
is older than a test specified version. However, it was discussed in the
WG that this should be a check for equality to avoid hitting issues
when a implementation tests against an older version of the CTS with
API breaking changes.

The tests for the command-buffer family of extensions are the only
provisional KHR tests using this versioning check, so this PR updates all
those cases to equality.