add initial unified SVM API tests (#2261)
This PR includes tests for `clSVMAllocWithPropertiesKHR` and
`clGetSVMPointerInfoKHR` for valid pointers, as per the unified SVM test
plan. These tests are passing on many devices with the unified SVM
emulation layer.

Note, these tests use the same test fixture and unified SVM wrapper as
the unified SVM capability tests in #2210.