Sanitize flags before passing them into clGetSupportedImageFormats
Some tests call get_format_list with e.g. CL_MEM_USE_HOSTPTR, but this
flag isn't valid clGetSupportedImageFormats, but it's not clear if
implementations are allowed to accept it.

This appears to be a problem since !2283 as now we get CL_MEM_USE_HOSTPTR
passed into get_format_list.