Remove semaphores_simple_2 and external_semaphores_simple_2 tests
While the cl_khr_semaphore extension spec does state that are no implicit
dependencies between already enqueued commands and the
clEnqueueSignalSemaphoresKHR, it's nothing special as this is already true
for any other event that's not a barrier or marker.

Also, the CTS can't reliably assume implementations to reorder events even
in an out of order queue as this is highly implementation defined behavior
and implementations may or may not choose to reorder events in a specific
order.

I don't see a reason why this should be tested for semaphores, but not for
any other commands, especially as it imposes a restriction on how to
implement out of order queues that wasn't enforced before.

Closes: https://github.com/KhronosGroup/OpenCL-CTS/issues/2439