Update `REQUIRE_EXTENSION` macro to use the new overload
Change `REQUIRE_EXTENSION` to use the new overload of
`is_extension_available`. The macro will now use its argument, `ext`,
to construct `ext##_EXTENSION_NAME` and `ext##_EXTENSION_VERSION`,
which, for example, correspond to `CL_KHR_SEMAPHORE_EXTENSION_NAME`
and `CL_KHR_SEMAPHORE_EXTENSION_VERSION`, respectively. The extension
name and version are both defined in the OpenCL headers.

This guarantees that the macro is checking for the latest extension
version before running the test, and eliminates a class of errors,
where the user has a spelling mistake in the extension name to
check. This is enforced at compile time, as a variable with a typo in
its name will not be defined in the headers.

Update all instances of `REQUIRE_EXTENSION` to match.

Add a new macro, `HAS_EXTENSION`, that also calls the new overload,
but does not check the result or return a test status if the call
failed. This is useful for instances where tests require acting upon
the result of the call, and not necessarily skip the test.

Signed-off-by: Ahmed Hesham <ahmed.hesham@arm.com>