Fix path to spv file in test_compiler.
In #1220, test_compiler was changed to load write_kernel.spv* from a
file using a relative path
test_conformance/compiler/spirv_bin/write_kernel.spv*. This works when
test_compiler is run from the build directory, but does not work when it
is run from the test_conformance directory, which is how
run_conformance.py expects to run it. Adjust the path accordingly.