Add helper to wrap result of std::filesystem::path::u8string()
In C++17 the return type of std::filesystem::path::u8string() is
std::string, but in C++20 the return type changed to std::u8string.

Add a helper to copy a std::u8string to a std::string, to be used
when a std::string is required. This fixes the build for C++20.