Test mutable command-buffer deferred kernel arguments
Tests the scenario outlined in
https://github.com/KhronosGroup/OpenCL-CTS/issues/2520#issuecomment-3280930492
to verify the functionality added in OpenCL-Docs PR
https://github.com/KhronosGroup/OpenCL-Docs/pull/1382.

Note that this bumps the required cl_khr_command_buffer_mutable_dispatch
minor version by 2 because
https://github.com/KhronosGroup/OpenCL-CTS/pull/2477 is not yet merged
which test the intermediate version.

This also assumes the proposed OpenCL-Docs change in
https://github.com/KhronosGroup/OpenCL-Docs/pull/1471 is merged, as
the PR does a negative test for command-buffer enqueue while the
command-buffer is in the finalized state.

Closes #2520